Introduce `@NonCacheable` annotation for Retrofit requests
This commit introduces a new annotation, `@NonCacheable`, which can be applied to Retrofit interface methods.

The `JsonDiskCacheInterceptor` has been updated to check for this annotation. If a request is marked as `@NonCacheable`, the interceptor will bypass the disk caching mechanism and proceed with the network request directly.

The `languages()` and `versions()` methods in `DataDragonApi` have been annotated with `@NonCacheable` as they are expected to change frequently and should not be cached.