Docs: Regenerate module dependency graphs
This commit updates the module dependency graphs to reflect recent changes.
The graphs are now generated with a newer version of the `modgraph` plugin (0.0.4) and have a slightly different visual style.

The following graphs have been updated:
- `app`
- `core`
- `core_data`
- `core_database`
- `core_datastore`
- `core_designsystemy`
- `core_model`
- `core_network`

The `provider` property in `generateModuleDependencyGraph` task in `build.gradle.kts` was removed as it's no longer necessary with the updated plugin.